requirements.txt
streamlit
SQLAlchemy>=2.0
psycopg[binary]>=3.1
python-dotenv>=1.0
pandas>=2.0